update media_file_size_bytes to BIGINT

+++ server/migrations/001-tables.js
@@ -34,7 +34,7 @@ exports.up = async (db) => {
             media_file_path TEXT, -- e.g. "/board/meetings/november.txt" path to the uploaded media at time of upload
             media_file_path_array TEXT[], -- same as media_file_path but split on path delimeters
             media_file_extension TEXT, -- upper-case file extension, e.g. MP4, PDF, etc
-            media_file_size_bytes INTEGER, -- size of object on S3
+            media_file_size_bytes BIGINT, -- size of object on S3
 
             -- media source info
             media_url TEXT, -- path to file location on S3
